make protocols to extension and add class file of table cell
